git_archive {gert}	R Documentation

Git Archive

Description

Exports the files in your repository to a zip file that is returned by the function.

Usage

git_archive_zip(file = NULL, repo = ".")

Arguments

file
name of the output zip file. Default is returned by the function

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

Value

path to the zip file that was created

See Also

Other git: git_branch(), git_commit(), git_config(), git_diff(), git_fetch(), git_merge(), git_rebase(), git_remote, git_repo, git_signature(), git_stash, git_tag

[Package gert version 1.5.0 Index]