git_open {gert}	R Documentation

Open local repository

Description

Returns a pointer to a libgit2 repository object.This function is mainly for internal use; users should simply reference a repository in gert by by the path to the directory.

Usage

git_open(repo = ".")

Arguments

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

Value

an pointer to the libgit2 repository

Examples

r <- tempfile(pattern = "gert")
git_init(r)
r_ptr <- git_open(r)
r_ptr
git_open(r_ptr)
git_info(r)

# cleanup
unlink(r, recursive = TRUE)

[Package gert version 1.5.0 Index]