git_remote {gert}	R Documentation

Git Remotes

Description

List, add, configure, or remove remotes.

Usage

git_remote_list(repo = ".")

git_remote_add(url, name = "origin", refspec = NULL, repo = ".")

git_remote_remove(remote, repo = ".")

git_remote_info(remote = NULL, repo = ".")

git_remote_set_url(url, remote = NULL, repo = ".")

git_remote_set_pushurl(url, remote = NULL, repo = ".")

git_remote_refspecs(remote = NULL, repo = ".")

Arguments

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

url
server url (https or ssh)

name
unique name for the new remote

refspec
optional string with the remote fetch value

remote
name of an existing remote. Default NULL means the remote from the upstream of the current branch.

See Also

Other git: git_archive, git_branch(), git_commit(), git_config(), git_diff(), git_fetch(), git_merge(), git_rebase(), git_repo, git_signature(), git_stash, git_tag

[Package gert version 1.5.0 Index]