git_reset {gert}	R Documentation

Reset your repo to a previous state

Description

git_reset_hard() resets the index and working tree

git_reset_soft() does not touch the index file or the working tree

git_reset_mixed() resets the index but not the working tree.

Usage

git_reset_hard(ref = "HEAD", repo = ".")

git_reset_soft(ref = "HEAD", repo = ".")

git_reset_mixed(ref = "HEAD", repo = ".")

Arguments

ref
string with a branch/tag/commit

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

See Also

Other git: git_archive, git_branch(), git_commit(), git_config(), git_diff(), git_fetch(), git_ignore, git_merge(), git_rebase(), git_remote, git_repo, git_signature(), git_stash, git_tag

[Package gert version 2.1.5 Index]