git_stash {gert}	R Documentation

Stashing changes

Description

Temporary stash away changed from the working directory.

Usage

git_stash_save(
  message = "",
  keep_index = FALSE,
  include_untracked = FALSE,
  include_ignored = FALSE,
  repo = "."
)

git_stash_pop(index = 0, repo = ".")

git_stash_drop(index = 0, repo = ".")

git_stash_list(repo = ".")

Arguments

message
optional message to store the stash

keep_index
changes already added to the index are left intact in the working directory

include_untracked
untracked files are also stashed and then cleaned up from the working directory

include_ignored
ignored files are also stashed and then cleaned up from the working directory

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

index
The position within the stash list. 0 points to the most recent stashed state.

See Also

Other git: git_archive, git_branch(), git_commit(), git_config(), git_diff(), git_fetch(), git_merge(), git_rebase(), git_remote, git_repo, git_signature(), git_tag

[Package gert version 1.5.0 Index]