git_tag {gert}	R Documentation

Git Tag

Description

Create and list tags.

Usage

git_tag_list(match = "*", repo = ".")

git_tag_create(name, message, ref = "HEAD", repo = ".")

git_tag_delete(name, repo = ".")

git_tag_push(name, ..., repo = ".")

Arguments

match
pattern to filter tags (use * for wildcard)

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

name
tag name

message
tag message

ref
target reference to tag

...
other arguments passed to git_push

See Also

Other git: git_archive, git_branch(), git_commit(), git_config(), git_diff(), git_fetch(), git_merge(), git_rebase(), git_remote, git_repo, git_signature(), git_stash

[Package gert version 1.5.0 Index]