user_is_configured {gert}	R Documentation

Test if a Git user is configured

Description

This function exists mostly to guard examples that rely on having a user configured, in order to make commits. user_is_configured() makes no distinction between local or global user config.

Usage

user_is_configured(repo = ".")

Arguments

repo
An optional repo, in the sense of git_open().

Value

TRUE if user.name and user.email are set locally or globally, FALSE otherwise.

Examples

user_is_configured()

[Package gert version 1.5.0 Index]