Git Branch

Create, list, and checkout branches.

Usage

git_branch(repo = ".")

git_branch_list(local = NULL, repo = ".")

git_branch_checkout(branch, force = FALSE, orphan = FALSE, repo = ".")

git_branch_create(
  branch,
  ref = "HEAD",
  checkout = TRUE,
  force = FALSE,
  repo = "."
)

git_branch_delete(branch, repo = ".")

git_branch_move(branch, new_branch, force = FALSE, repo = ".")

git_branch_fast_forward(ref, repo = ".")

git_branch_set_upstream(upstream, branch = git_branch(repo), repo = ".")

git_branch_exists(branch, local = TRUE, repo = ".")

Arguments

repo

The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

local

set TRUE to only check for local branches, FALSE to check for remote branches. Use NULL to return all branches.

branch

name of branch to check out

force

overwrite existing branch

orphan

if branch does not exist, checkout unborn branch

ref

string with a branch/tag/commit

checkout

move HEAD to the newly created branch

new_branch

target name of the branch once the move is performed; this name is validated for consistency.

upstream

remote branch from git_branch_list, for example "origin/master"

See also