git_diff {gert}	R Documentation

Git Diff

Description

View changes in a commit or in the current working directory.

Usage

git_diff(ref = NULL, repo = ".")

git_diff_patch(ref = NULL, repo = ".")

Arguments

ref
a reference such as "HEAD", or a commit id, or NULL to the diff the working directory against the repository index.

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

See Also

Other git: git_archive, git_branch(), git_commit(), git_config(), git_fetch(), git_merge(), git_rebase(), git_remote, git_repo, git_signature(), git_stash, git_tag

[Package gert version 1.5.0 Index]