git_ignore {gert}	R Documentation

Git Ignore

Description

Test if files would be ignored by .gitignore rules

Usage

git_ignore_path_is_ignored(path, repo = ".")

Arguments

path
A character vector of paths to test within the repo

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

Value

A logical vector the same length as path, indicating if the paths would be ignored.

See Also

Other git: git_archive, git_branch(), git_commit(), git_config(), git_diff(), git_fetch(), git_merge(), git_rebase(), git_remote, git_repo, git_reset(), git_signature(), git_stash, git_tag

[Package gert version 2.1.5 Index]