git_checkout_pull_request {gert}	R Documentation

GitHub Wrappers

Description

Fetch and checkout pull requests.

Usage

git_checkout_pull_request(pr = 1, remote = NULL, repo = ".")

git_fetch_pull_requests(pr = "*", remote = NULL, repo = ".")

Arguments

pr
number with PR to fetch or check out. Use "*" to fetch all pull requests.

remote
Optional. Name of a remote listed in git_remote_list(). If unspecified and the current branch is already tracking branch a remote branch, that remote is honored. Otherwise, defaults to origin.

repo
The path to the git repository. If the directory is not a repository, parent directories are considered (see git_find). To disable this search, provide the filepath protected with I(). When using this parameter, always explicitly call by name (i.e. repo = ) because future versions of gert may have additional parameters.

Details

By default git_fetch_pull_requests will download all PR branches. To remove these again simply use git_fetch(prune = TRUE).

[Package gert version 1.5.0 Index]